backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
groq>=0.9.0
PyMuPDF==1.24.5
python-docx==1.1.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
openai-whisper==20231117
numpy==1.26.4
Pillow==10.4.0
werkzeug==3.0.3
gunicorn==22.0.0
